+++ src/Actions/Action.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Binarcode\LaravelMailator\Actions;
+
+use Binarcode\LaravelMailator\Models\MailatorSchedule;
+
+interface Action
+{
+    public function handle(MailatorSchedule $schedule);
+}